+++ sample-basic-java/src/main/java/org/dbtools/sample/model/database/DatabaseBaseManager.java
@@ -33,6 +33,7 @@ public abstract class DatabaseBaseManager extends AndroidDatabaseManager {
         
         // Tables
         AndroidBaseManager.createTable(database, org.dbtools.sample.model.database.main.individual.IndividualConst.CREATE_TABLE);
+        AndroidBaseManager.createTable(database, org.dbtools.sample.model.database.main.individualdata.IndividualDataConst.CREATE_TABLE);
         
         database.setTransactionSuccessful();
         database.endTransaction();