+++ sample-basic-kotlin/src/main/kotlin/org/dbtools/sample/kotlin/model/database/main/individualdata/IndividualDataManager.kt
@@ -1,0 +1,22 @@
+/*
+ * IndividualDataManager.kt
+ *
+ * Generated on: 11/20/2016 11:05:51
+ *
+ */
+
+
+
+package org.dbtools.sample.kotlin.model.database.main.individualdata
+
+import org.dbtools.sample.kotlin.model.database.DatabaseManager
+
+
+class IndividualDataManager : IndividualDataBaseManager {
+
+
+    constructor(databaseManager: DatabaseManager) : super(databaseManager) {
+    }
+
+
+}