+++ sample-rxjava/src/main/java/org/dbtools/sample/model/database/main/individualdata/IndividualData.java
@@ -1,0 +1,34 @@
+/*
+ * IndividualData.java
+ *
+ * Created: 11/20/2016 11:05:19
+ */
+
+
+
+package org.dbtools.sample.model.database.main.individualdata;
+
+import android.database.Cursor;
+import org.dbtools.android.domain.database.contentvalues.DBToolsContentValues;
+
+
+public class IndividualData extends IndividualDataBaseRecord {
+
+
+    public IndividualData(IndividualData record) {
+        super(record);
+    }
+
+    public IndividualData(Cursor cursor) {
+        setContent(cursor);
+    }
+
+    public IndividualData(DBToolsContentValues values) {
+        setContent(values);
+    }
+
+    public IndividualData() {
+    }
+
+
+}