+++ src/main/java/org/dbtools/sample/model/database/main/individual/IndividualManager.java
@@ -1,0 +1,23 @@
+/*
+ * IndividualManager.java
+ *
+ * Generated on: 11/20/2016 10:26:47
+ *
+ */
+
+
+
+package org.dbtools.sample.model.database.main.individual;
+
+import org.dbtools.sample.model.database.DatabaseManager;
+
+
+public class IndividualManager extends IndividualBaseManager {
+
+
+    public IndividualManager(DatabaseManager databaseManager) {
+        super(databaseManager);
+    }
+
+
+}