+++ src/main/java/org/dbtools/sample/model/database/main/individualdata/IndividualDataManager.java
@@ -1,0 +1,23 @@
+/*
+ * IndividualDataManager.java
+ *
+ * Generated on: 11/20/2016 10:26:47
+ *
+ */
+
+
+
+package org.dbtools.sample.model.database.main.individualdata;
+
+import org.dbtools.sample.model.database.DatabaseManager;
+
+
+public class IndividualDataManager extends IndividualDataBaseManager {
+
+
+    public IndividualDataManager(DatabaseManager databaseManager) {
+        super(databaseManager);
+    }
+
+
+}