+++ src/main/java/org/dbtools/sample/model/database/main/individualview/IndividualViewManager.java
@@ -1,0 +1,23 @@
+/*
+ * IndividualViewManager.java
+ *
+ * Generated on: 11/20/2016 10:26:47
+ *
+ */
+
+
+
+package org.dbtools.sample.model.database.main.individualview;
+
+import org.dbtools.sample.model.database.DatabaseManager;
+
+
+public class IndividualViewManager extends IndividualViewBaseManager {
+
+
+    public IndividualViewManager(DatabaseManager databaseManager) {
+        super(databaseManager);
+    }
+
+
+}